Crate pica_matcher

This crate provides various matcher to filter PICA+ records, fields or subfields.

Re-exports

pub use field_matcher::FieldMatcher;

pub use subfield_matcher::SubfieldMatcher;

Modules

field_matcher

Matcher that works on PICA+ Fields.

parser

Parsers recognizing matcher for PICA+ primitves.

subfield_matcher

Matcher that works on PICA+ Subfields.

Structs

MatcherOptions

Options and flags which can be used to configure a matcher.

RecordMatcher

A Matcher that works on PICA+ Records.

Enums

OccurrenceMatcher

A matcher that matches against PICA+ Occurrence.

ParseMatcherError

An error that can occur when parsing matcher expressions.

TagMatcher

A matcher that matches against PICA+ Tags.