Enum pica_matcher::field_matcher::FieldMatcher

pub enum FieldMatcher {
    Singleton(SingletonMatcher),
    Cardinality(CardinalityMatcher),
    Group(Box<FieldMatcher>),
    Not(Box<FieldMatcher>),
    Composite {
        lhs: Box<FieldMatcher>,
        op: BooleanOp,
        rhs: Box<FieldMatcher>,
    },
}

A field matcher that allows grouping, negation and connecting of singleton matcher.

Variants

Singleton(SingletonMatcher)

Cardinality(CardinalityMatcher)

Group(Box<FieldMatcher>)

Not(Box<FieldMatcher>)

Composite

Fields

lhs: Box<FieldMatcher>

op: BooleanOp

rhs: Box<FieldMatcher>

Implementations

impl FieldMatcher

pub fn new(data: &str) -> Result<Self, ParseMatcherError>

Create a new field matcher from a string slice.

Example

use pica_matcher::FieldMatcher;
use pica_record::FieldRef;

fn example() -> anyhow::Result<()> {
    let matcher = FieldMatcher::new("003@?")?;

    assert!(matcher.is_match(
        &FieldRef::new("003@", None, vec![('0', "123456789X")]),
        &Default::default()
    ));

    Ok(())
}

pub fn is_match<'a, T: AsRef<[u8]> + 'a>(
&self,
fields: impl IntoIterator<Item = &'a Field<T>> + Clone,
options: &MatcherOptions
) -> bool

Returns true if the given field matches against the field matcher.