Struct pica_matcher::RecordMatcher

pub struct RecordMatcher { /* private fields */ }

A Matcher that works on PICA+ Records.

Implementations

impl RecordMatcher

pub fn new<T: ?Sized + AsRef<[u8]>>(data: &T) -> Self

Create a new field matcher from a string slice.

Example

use pica_matcher::RecordMatcher;
use pica_record::RecordRef;

fn example() -> anyhow::Result<()> {
    let matcher = RecordMatcher::new("003@?");
    let record =
        RecordRef::new(vec![("003@", None, vec![('0', "abc")])]);

    assert!(matcher.is_match(&record, &Default::default()));
    Ok(())
}

pub fn is_match(&self, record: &RecordRef<'_>, options: &MatcherOptions) -> bool

Returns true if the given record matches against the record matcher.