Module pica_matcher::subfield_matcher

Matcher that works on PICA+ Subfields.

Structs

CardinalityMatcher

A matcher that checks the number of occurrences of a subfield.

ExistsMatcher

A matcher that checks if a subfield exists.

InMatcher

A matcher that checks if a subfield value is in a predefined list.

RegexMatcher

A matcher that checks a subfield value against a regex.

RelationMatcher

A matcher that checks relations between (string) values.

Enums

SingletonMatcher

A matcher that checks for the singleton matcher.

SubfieldMatcher

A matcher that allows grouping, negation and connecting of singleton matcher.

Constants

GROUP_LEVEL

Functions

parse_subfield_matcher